% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cross_sectional_logistics}
\alias{cross_sectional_logistics}
\title{Cross-sectional Study Logistics}
\usage{
cross_sectional_logistics(
  final_n,
  rejection_rate,
  eligibility_rate,
  subjects_per_day,
  working_days_month
)
}
\arguments{
\item{final_n}{Final calculated sample size}

\item{rejection_rate}{Expected rejection rate (proportion between 0 and 1)}

\item{eligibility_rate}{Expected eligibility rate (proportion between 0 and 1)}

\item{subjects_per_day}{Number of subjects/records that can be processed per day}

\item{working_days_month}{Number of working days per month}
}
\value{
A list with study logistics calculations
}
\description{
Cross-sectional Study Logistics
}
